bump version, ant example
