+++ examples/ant/package.json
@@ -3,10 +3,11 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@enterprise-oss/osso": "^0.0.8",
+    "@enterprise-oss/osso": "0.0.9",
     "antd": "^4.4.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1"
   },
   "scripts": {
