+++ examples/ant/src/index.js
@@ -1,21 +1,15 @@
 import './index.css';
 import 'antd/dist/antd.css';
 
-import { OssoProvider } from '@enterprise-oss/osso';
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { BrowserRouter as Router } from 'react-router-dom';
 
 import App from './App';
 
 ReactDOM.render(
-  <OssoProvider
-    client={{
-      baseUrl: 'https://demo.ossoapp.com',
-      jwt:
-        'eyJhbGciOiJIUzI1NiJ9.eyJlbWFpbCI6InNhbUBhZG1pbi5jb20iLCJzY29wZSI6ImFkbWluIiwiaWQiOiI2YjJmYzNkMS0wYjU4LTQ0ZTYtYWZkNC04YzMzMzg4NzIxZjYiLCJvYXV0aF9jbGllbnRfaWQiOiI0MmZiMzFkOGZiYzYyYWY2MzgzN2RiNTNhNzkwOGU4MSJ9.HInHbrbtNjLykPPRZiJMOuzu3St0SSmeYoPC2fUqLVo',
-    }}
-  >
+  <Router>
     <App />
-  </OssoProvider>,
+  </Router>,
   document.getElementById('root'),
 );
