+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@enterprise-oss/osso",
   "license": "SEE LICENSE IN LICENSE",
-  "version": "0.0.8",
+  "version": "0.0.9",
   "description": "React components for Osso",
   "main": "dist/index.js",
   "module": "dist/index.esm.js",
